requirements.txt
amqp==5.2.0
asgiref==3.8.1
asn1crypto==1.5.1
babel==2.16.0
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorclass==2.2.2
cryptography==43.0.0
Django==5.1
django-redis==5.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
docopt==0.6.2
flower==2.0.1
humanize==4.10.0
idna==3.8
kombu==5.4.0
packaging==24.1
pip-upgrader==1.4.15
prometheus_client==0.20.0
prompt_toolkit==3.0.47
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2018.5
redis==5.0.8
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
terminaltables==3.1.10
tornado==6.4.1
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
